[Calc] Removing identifier logic that relies on IP to SG link (Recidiviz/recidiviz-data#10042)
This removes the fields most_serious_offense_ncic_code and most_serious_offense_statute from the IncarcerationPopulationMetric. These fields are entirely unused, and rely on the schema link between incarceration periods and sentence groups. Deleting this in preparation for removing this schema link when we move periods onto the person.

GitOrigin-RevId: db4e367109f0a3081a2a65ed65d5e66680fd9be8