[US_TN] Adding `custodial_authority` to IPs (Recidiviz/recidiviz-data#10292)
Brings in the SiteType column from the Site table that determines the type of a facility. This value can be used to determine custodial authority in US_TN, so this PR adds that to the ingest mapping. Will be used to only include STATE_PRISON periods in the metrics for the state's population.

GitOrigin-RevId: 3d212872afbd6ac5bac92400dd270417ba851fd4