[Entity matching] Merge certain entities atomically (Recidiviz/recidiviz-data#10518)
This change makes it so that when most of the entities we ingest are merged in entity matching, we just always pull ALL the values from the new entity so values don't get mangled together.

This does NOT address the issue where raw text and enum values are not merged atomically for entities that aren't merged fully atomically.

GitOrigin-RevId: eaff711e99ca24fdf24601532b95bde564f72777