[GCS] Use rewrite instead of copy_blob for file copies (Recidiviz/recidiviz-data#10751)
In the process of trying to clean up the TN raw data bucket in prod, I wrote a small hacky script to copy a bunch of the raw data files in the ingest bucket to storage. The script kept failing with the "Copy spanning locations and/or storage classes could not complete within 30 seconds" error. With this fix, copies of large files do not fail.

Closes Recidiviz/recidiviz-data#4286

GitOrigin-RevId: 6188eca4b1f536c96eab69ae2b7851753f49a78e