[Normalization] Bidirectional entity updates (Recidiviz/recidiviz-data#10859)
Adds a new function `deep_entity_update` that is now used throughout entity normalization when we want to make an update to an entity. The function builds a new version of the entity with the updated attributes, and for any attributes that are references to other entities updates the reverse references on those entities to point to the new, updated version of the original entity.

Added a new lint rule to make sure we don't use `attr.evolve` in any `entity_normalization` file.

GitOrigin-RevId: 13ccfa8d6cad3ffc8d12a0160497122de2c1faff