[Deploy script] Updating serving versions check (Recidiviz/recidiviz-data#11399)
We have been running into issues deploying recently where we hit the BACKEND_SERVICES quota at the end of the deploy. We have a backend service in our projects that is distinct from our serving versions, and our check_for_too_many_serving_versions check wasn't accounting for this.

Bumps the max number of serving versions down by 1 to account for it.

GitOrigin-RevId: db698193f9036c268ceb58bfbd1ec5c969aca559