[BigQuery views] New shared_metric_views dataset (Recidiviz/recidiviz-data#11406)
We are going to move the view update part of the deploy to happen after the app is deployed. Until we are able to put the view update in the DAG, we will need to update the views at at the same time as triggering the Dataflow pipelines after the deploy. There are reference views that the pipelines rely on that will need to get updated as part of the deploy. In order to make that process as fast as possible, this PR moves all views from the reference_views dataset that are descendants of dataflow metrics into a new shared_metric_views dataset.

This adds tests to enforce that none of the reference_views rely on dataflow metrics and that all of the shared_metric_views are descendants from dataflow metrics.

GitOrigin-RevId: b208dbcb959113e344666b9e18aefb934bce9f46