[BigQuery View Deploy] Ability to deploy subgraph (Recidiviz/recidiviz-data#11445)
Adds the ability to deploy just a subgraph of views. We will need this because we need to deploy the reference views during the app deploy when we move the rest of view deploy/materialization to post-deploy. This is because the dataflow jobs rely on reference views, so those need to be updated definitively before any Dataflow jobs run.

GitOrigin-RevId: 194c2d9cc9157e20ae67cfcd4b7d7c8d392f5b85