[BQ materialization] Refactor export manager to isolate file-based materialization logic (Recidiviz/recidiviz-data#11367)
Currently, the `DirectIngestIngestViewExportManager`, which holds logic related to exporting ingest view results to GCS has two main high-level roles:
1) Generate args for new ingest view export (materialization) tasks
2) Actually perform the export for a given set of args

This PR splits the class into two, keeping the export logic in the `DirectIngestIngestViewExportManager`, but moving the args generation logic to a new implementation of a new interface, `IngestViewMaterializationTaskArgsGenerator`.

For both classes, added a new delegate object that isolates the details of file-based materialization behind an interface.

Pre-work for Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: 307154db698bdb6968fbfa574d5d6902242f3ded