[Entity Normalization] Unique IDs on normalized entities (Recidiviz/recidiviz-data#11477)
Fixes a few issues we were seeing with entity normalization in US_MO and US_ND. These are both states where we create new entities, where those new entities are connected to related entities that existed before. In US_MO we create new `StateSupervisionPeriod` entities, and hydrate them with existing `StateSupervisionCaseTypeEntry` entities. In US_ND we create new `StateSupervisionViolation` and hydrate them with existing `StateSupervisionViolationTypeEntry` entities.

This PR does the following:
- Creates a mechanism for creating ID values that will be unique across all normalized entities of a given type
- Adds a `copy_entities_and_add_unique_ids` function that is used when putting existing entities onto a new entity (so that we don't try to re-use child entities when hydrating a new entity)

An assumption made in this approach: Caching the entity IDs that have been used for a given `person_id` relies on the assumption that all of the work in the `normalize_entities` step for a single `person_id` will happen on a single thread on a given worker, because if we lose the global cache at any point mid-processing we will lose the IDs that have already been used. This is not guaranteed by Dataflow, but I have found it true in practice with our data and our pipelines. I filed [this issue](https://github.com/Recidiviz/recidiviz-data/issues/11470) to write validations that will fail if this assumption stops being true.

GitOrigin-RevId: 60c824327979c97af9e68861c3f04bbd3407a1a9