[Entity Normalization] Using Normalized IPs in identifiers (Recidiviz/recidiviz-data#11541)
This updates the IP normalization manager to return `Tuple[List[StateIncarcerationPeriod], AdditionalAttributesMap]` instead of `NormalizedIncarcerationPeriodIndex`, and updates the NormalizedIncarcerationPeriodIndex to store the Normalized versions of IPs (and to sort on sequence_num on instantiation). Adds a temporary step in the identifiers to convert the processed IPs into their Normalized versions to instantiate the NormalizedIncarcerationPeriodIndex, and updates all downstream functions to use NormalizedStateIncarcerationPeriod entities.

GitOrigin-RevId: cc40b1b1c0ca902a96735fb5f2ef1dec45cd76fa