[Testing] Refactor FakeDirectIngestBigQueryClient to generalize (Recidiviz/recidiviz-data#11699)
Renames the `FakeDirectIngestBigQueryClient` to `FakeBigQueryClient` and moves it to `recidiviz.tests.big_query.fakes`. Eliminates all usages of the `FakeBigQueryClient` by mocking out the `IngestViewMaterializer` at a higher level for tests in `base_direct_ingest_controller_test.py`.

In a follow-up PR, will update the `FakeBigQueryClient` to use a `FakeBigQueryDatabase` so it can actually mock behavior using Postgres. I'll use that class for testing lower-level classes that talk to BigQuery.

Related to Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: 0ccfa29745159f421fadadedee6d6f7a3fb8f94d