[Bugfix] Update Data Freshness view to not encorporate data from secondary (Recidiviz/recidiviz-data#11871)
The "ingest_metadata_latest_ingested_upper_bounds" view backs the Data Freshness tab of the admin panel. That page is supposed to reflect the freshness of the data in the BigQuery state dataset. That data is only reading from data in the PRIMARY Postgres database, but did not filter to only PRIMARY rows in the metadata.

Closes Recidiviz/recidiviz-data#10664

GitOrigin-RevId: befac6ce4474b30cd53d87fb1f9faed031949097