[BQ materialization] Add interface for accessing ingest view results in BQ (Recidiviz/recidiviz-data#11836)
Adds a new InstanceIngestViewContents which provides an interface for writing to / reading from ingest view results tables in BigQuery.

I spent a bunch of time trying to test this against the FakeBigQueryDatabase, but there were a lot of nuances which made it extremely slow going and left me not particularly confident that I was testing what I was intending to test.

Related to Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: b50e0c106391faf1a6a86db5c7d88ab807523fac