[BQ materialization] Use InstanceIngestViewContents in IngestViewMaterializer (Recidiviz/recidiviz-data#11857)
Integrates the new InstanceIngestViewContents into the BQBasedMaterializerDelegate used by the IngestViewMaterializer.

The vast majority of the changes here are in tests. I copied all the IngestViewMaterializer tests that use the legacy file-based classes and adjusted to use the BQ-based ones. The old tests will be deleted once we ship to all states.

Related to Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: f7b47828fe4c1772be384cbcdada0efddc506c04