[Justice Counts] Add name and email to UserAccount table (Recidiviz/recidiviz-data#11902)
Originally I thought we'd keep email addresses and names in Auth0, and that there wasn't a need to store another copy of them in our database. But @terryttsai made the really good observation that we need to be able to create users manually before they sign in via Auth0, so that we can add them to the right agencies in advance, and avoid a bad user experience where the user isn't connected to any agencies upon login. So here we augment the UserAccount model with email address (required) and name (optional) and make auth0 user id optional.

GitOrigin-RevId: 29320820cb523dbd8a6e8031ccf881302140115b