[Justice Counts][1/n] Adding `ReportedMetric` class to represent metric values reported by agencies (Recidiviz/recidiviz-data#11844)
The main contribution of this PR is to create a new ReportedMetric class, which represents an agency's reported values for an official Tier 1 Justice Counts metric. There will be a many:1 mapping between instances of this class and instances of MetricDefinition, since agencies will report the same metric over different time periods on a recurring basis.

GitOrigin-RevId: 82ae09e7ffc59ec3a14edacd7874a5dd2e1e5d60