[BigQuery Views] Fixing bad dataset ID raw values in queries (Recidiviz/recidiviz-data#12043)
Adds a new test that validates that all dataset ID values are passed in as arguments to the view builder, and are not in the raw query template. Required a bunch of view queries to be updated.

There were no updates to any of the view docs, which is a good sign that there were no functional changes to any of the view queries (aka none of the dependency graphs changed for any of the views). Did a test view deploy and all views deployed successfully.

GitOrigin-RevId: 2662cba423a69ac227f67ea95835d676fd56e2b5