[Justice Counts] Script to make backend API testing easier (Recidiviz/recidiviz-data#12142)
Introduces a tool to make backend testing of JC APIs easier (i.e. no frontend required!). You just specify the endpoint you want and the payload for it, and the tool will handle all authentication, make the request, and print out the response. See the update to the README for instructions on how to use.

GitOrigin-RevId: 216e764d72f0805fc3d0e28531ab3d12a474d3f8