[Enums] Moving ChargeStatus to StateChargeStatus (Recidiviz/recidiviz-data#12208)
We need to add new features to the EntityEnum class so that we can build enum docs (see PR Recidiviz/recidiviz-data#12176). We need to be able to add features to the enums used in the state schema without interfering with the county schema. This requires killing the notion of shared_enums between schemas.

This PR adds a new StateChargeStatus enum and migrates all of the state uses of ChargeStatus to use StateChargeStatus instead.

GitOrigin-RevId: 10a959968a3ff98df934df646f747ed877be3c1e