[Normalization/Validations] Updating normalized entity validations (Recidiviz/recidiviz-data#12153)
Updates a bunch of validations that alert to problems in IP normalization logic to pull from the normalized_state dataset instead of the metrics. Also adds the normalized_state dataset to the two "freshness" validations.

Confirmed that validation output was the same. Had to update a few validations to handle TRANSFER admission/release reasons, since these are excluded from metrics but are present in normalized IPs.

GitOrigin-RevId: 1457c5bb0342425ac7d3496b7c46e5219148c876