[Postgres deprecation] Delete v1 Postgres instances (Recidiviz/recidiviz-data#12033)
This removes the instances that we can delete now that we've passed the 6-month database backup retention period, for those database instances that were on Postgresql versions earlier than 13. Confirmed these DBs are no longer being used and can be safely deleted.

Out of an abundance of caution, will land this in a separate prod release from Recidiviz/recidiviz-data#12031, so will not land until after the staging deploy has been cut on Friday 4/15.

Part of Recidiviz/recidiviz-data#11326

GitOrigin-RevId: e9ab79e7b82b0aa94be10b11d4db10d50eb995c0