[Terraform] Default deletion_protection to False for all BQ tables and CloudSQL instances (Recidiviz/recidiviz-data#12329)
Had a discussion with @danawillow in which I learned that the deletion_protection flag was added as a default to a few google TF provider objects so that it was hard for people to delete data sources if they were using CI or some automatic process to apply TF. Since we run and inspect TF manually, this flag is more of a nuisance than anything else. Setting the default to false for all BQ tables and CloudSQL instances.

Related to Recidiviz/recidiviz-data#11326

GitOrigin-RevId: 0f7ba70372bf7a2876aabf47bfcbb3283f786091