[Docs] Enum documentation support (Recidiviz/recidiviz-data#12176)
Adds support for enum documentation, with enforcements that the documentation will stay up-to-date.

Moves all of the entity docs from docs/schema/ into docs/schema/entities. Creates a new docs/schema/enums folder with all of the enum doc files. Adds links from the field in an entity that stores an enum to the doc for that enum class.

GitOrigin-RevId: e562cb9d06421acf62c019805bfe622fb5337870