[Ingest] Expose operations DB processed file counts in admin panel API (Recidiviz/recidiviz-data#12494)
We currently show counts of operations DB rows for unprocessed raw / ingest view files in the admin panel, but not counts of processed files. Adds functionality to count processed files and pipes it through to the admin panel API.

Note: all functionality related to ingest view files will go away soon, but this was easy to do and nice for symmetry.

Related to Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: af1e2447bef4e41a746b103e1d0b6355dfc36dbf