[View Deploy] Fixing view deploy endpoint trigger (Recidiviz/recidiviz-data#12556)
Fixes the bug that was not including the update_managed_views request arg in the body of the request that hit the /refresh_bq_schema endpoint, causing the view update endpoint to not get hit.

This does not remove the view update out of the deploy. We will get a PR in that does that as soon as we've confirmed that this is fully working. Views will be double-deployed until next PR is deployed that removes the view deploy from the deploy.

GitOrigin-RevId: 051239e39c88bc56bce01eeebe1f522b552862e3