[Admin Panel] Parallelize ingest instance info fetch (Recidiviz/recidiviz-data#12620)
In order to fix Recidiviz/recidiviz-data#12600, we will need to start doing a much more expensive GCS call to scan the storage bucket for processed files. Adding this simple change before I do that, which just parallelizes some of the loading work that the ingest ops page does, so that we now load data for each instance in parallel.

Prep for Recidiviz/recidiviz-data#12600.

GitOrigin-RevId: 0b7491029b16ebd980293a67a49e5b275b3287b7