[US_ND] Remove bad columns from docstars_offenders mappings (Recidiviz/recidiviz-data#12739)
When I migrated `docstars_offenders` to v2 mappings, I blindly copied the columns in the `keys_to_ignore` section of the v1 mappings to the `unused_columns` section of v2 mappings. However, some of the columns in `keys_to_ignore` are not actually present in the raw data (they may have been present in a very early version of the data, which is why they were listed there?).

The v2 mappings are more strict and will crash if there are columns missing from a results row but listed as an input col, so we were crashing with:
```
ValueError: Found columns in manifest |input_columns| list that are missing from results row [0]: {'LAST_UPDATE', 'PHONE2', 'ABSDATE', 'WEIGHT', 'HEIGHT'}
```

This will need to be cherry-picked to prod.

GitOrigin-RevId: 92cf1e2c7e7ba13d3e199e82177bed2a1886110b