[BQ materialization] Do not crash Data Freshness page if BQ materialization is enabled (Recidiviz/recidiviz-data#12763)
I am starting to migrate MI to BQ materialization and I want to be able to test flashing SECONDARY -> PRIMARY but it will cause the whole freshness page to crash when I do. @elchao96 told me it's ok if MI does not have good data on that page, so I updated logic to show a soft error.

Related to Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: d000fd63ca681c39b22ee6fa3e3947635513ba4e