[BQ materialization] Perf optimization for extract and merge job prioritizer (Recidiviz/recidiviz-data#12804)
Changes the structure of get_next_unprocessed_batch_info to query for the highest priority batch for all view tables at once.

This means that even if we have 100 ingest view queries for a state, we only do one BQ query in the /scheduler endpoint.

Perf improvement for Recidiviz/recidiviz-data#9717.

GitOrigin-RevId: 8114809c44ed09c2d3cdb8f93102919839bccc1a